Duplicating color_ramp and curve on creation so that user doesn't change the default resource by accident
On creation, GPUTrail already fills the color_ramp and curve with the default resources. However, it keeps a reference to the original resource instead of creating a new one for the new node. This means that, unless the user changes those resources manually by marking them as unique or assigning a new one, the default resource will be changed which might impact other instances of that node unintentionally